feat: added app user and custom contacts classes and basic endpoints

+++ api/views.py
@@ -7,8 +7,8 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import status
 
-from api.models import Task, Subtask
-from join_backend.serializers import TaskSerializer, SubtaskSerializer
+from api.models import AppUser, CustomContact, Task, Subtask
+from join_backend.serializers import AppUserSerializer, CustomContactSerializer, TaskSerializer, SubtaskSerializer
 
 
 class LoginView(ObtainAuthToken):
@@ -27,6 +27,7 @@ class LoginView(ObtainAuthToken):
 
 class RegisterView(ObtainAuthToken):
     # POST
+    # ALSO CREATE APPUSER INSTANCE
     pass
 
 
@@ -61,9 +62,11 @@ class TasksView(APIView):
 
     def delete(self, request, *args, **kwargs):
         pk = kwargs.get('pk')
-        task = Task.objects.get(id=pk)
-        task.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if pk:
+            task = Task.objects.get(id=pk)
+            task.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
     
     
 class SubtasksView(APIView):
@@ -97,16 +100,63 @@ class SubtasksView(APIView):
 
     def delete(self, request, *args, **kwargs):
         pk = kwargs.get('pk')
-        subtask = Subtask.objects.get(id=pk)
-        subtask.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if pk:
+            subtask = Subtask.objects.get(id=pk)
+            subtask.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
     
     
 class UsersView(APIView):
-    # GET, POST
-    pass
+    authentication_classes = [TokenAuthentication]
+    permission_classes = [IsAuthenticated]
+    
+    
+    def get(self, request, format=None):
+        users = AppUser.objects.all()
+        serializer = AppUserSerializer(users, many=True)
+        return Response(serializer.data)
     
     
 class ContactsView(APIView):
-    # GET, POST, DELETE, PUT
-    pass
+    authentication_classes = [TokenAuthentication]
+    permission_classes = [IsAuthenticated]
+    
+    
+    def get(self, request, format=None):
+        custom_contacts = CustomContact.objects.filter(user=request.user)
+        serializer = CustomContactSerializer(custom_contacts, many=True)
+        return Response(serializer.data)
+
+
+    def post(self, request, format=None):
+        serializer = CustomContactSerializer(data=request.data)
+        if serializer.is_valid():
+            if int(request.data['user']) != request.user.id:
+                return Response(status=status.HTTP_403_FORBIDDEN)                
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+
+    def put(self, request, *args, **kwargs):
+        pk = kwargs.get('pk')
+        custom_contact = CustomContact.objects.get(id=pk)
+        if custom_contact.user != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+        serializer = CustomContactSerializer(custom_contact, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+    def delete(self, request, *args, **kwargs):
+        pk = kwargs.get('pk')
+        if not pk:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        custom_contact = CustomContact.objects.get(id=pk)
+        if custom_contact.user != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+        custom_contact.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
